+++ alignment/system3.py
@@ -1,17 +1,14 @@
-
 import json
-import scipy.stats
-import numpy as np
 import os
-
 from copy import deepcopy
-from loggers import logger
+
 import matplotlib.pyplot as plt
+import numpy as np
+import scipy.stats
 
+from submission.loggers import logger
+from submission.parsers import System3LogParser
 from system1 import UnAlignableEEGException
-from parsers.system3_log_parser import System3LogParser
-
-from configuration import config
 
 
 class System3Aligner(object):
