+++ submission/parsers/electrode_config_parser.py
