+++ submission/parsers/ltpfr2_log_parser.py
@@ -1,4 +1,4 @@
-from base_log_parser import BaseSessionLogParser
+from .base_log_parser import BaseSessionLogParser
 import numpy as np
