+++ tests/test_event_creation.py
@@ -1,11 +1,12 @@
+import os
 
+import numpy as np
+from parsers.base_log_parser import StimComparator
+from parsers.catfr_log_parser import CatFRSessionLogParser
 from parsers.fr_log_parser import FRSessionLogParser
 from parsers.pal_log_parser import PALSessionLogParser
-from parsers.math_parser import MathLogParser
-from parsers.catfr_log_parser import CatFRSessionLogParser
-import numpy as np
-import os
-from parsers.base_log_parser import StimComparator
+
+from submission.parsers.math_parser import MathLogParser
 
 DATA_ROOT = '/Volumes/rhino_mount/data/eeg/'
